+++ Bonni/AttentiveExample/CreateAccountViewController.swift
@@ -69,7 +69,7 @@ class CreateAccountViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .white
+        view.backgroundColor = .systemBackground
         title = "Create an account"
         setupUI()
     }