+++ Bonni/AttentiveExample/DebugConsoleViewController.swift
@@ -20,7 +20,7 @@ class DebugConsoleViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .white
+        view.backgroundColor = .systemBackground
 
         view.alpha = 0.8