+++ Bonni/AttentiveExample/PlaceOrderViewController.swift
@@ -86,7 +86,7 @@ class PlaceOrderViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .white
+        view.backgroundColor = .systemBackground
         title = "Place Order"
         setupUI()
     }